+++ test/libsolidity/smtCheckerTests/array_members/pop_constructor_unsafe.sol
@@ -1,0 +1,10 @@
+pragma experimental SMTChecker;
+
+contract C {
+	uint[] a;
+	constructor() public {
+		a.pop();
+	}
+}
+// ----
+// Warning: (83-90): Empty array "pop" detected here.