+++ Changelog.md
@@ -21,10 +21,14 @@ Language Features:
 
 
 Compiler Features:
+ * Build system: Update the soljson.js build to emscripten 1.39.15 and boost 1.73.0 and include Z3 for integrated SMTChecker support without the callback mechanism.
+ * SMTChecker: Support array ``length``.
+ * SMTChecker: Support array ``push`` and ``pop``.
 
 
 Bugfixes:
-
+ * Optimizer: Fixed a bug in BlockDeDuplicator.
+ * Type Checker: Disallow assignments to storage variables of type ``mapping``.
 
 
 ### 0.6.8 (2020-05-14)